% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inNout.R
\name{textillate_in}
\alias{textillate_in}
\alias{textillate_out}
\title{Customise effects}
\usage{
textillate_in(t, effect = "fadeInLeftBig", delay.scale = 1.5, delay = 50,
  sync = FALSE, shuffle = FALSE, reverse = FALSE, ...)

textillate_out(t, effect = "hinge", delay.scale = 1.5, delay = 50,
  sync = FALSE, shuffle = FALSE, reverse = FALSE, ...)
}
\description{
Cutomise in  and out effect.
}
