% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{textillateStartProxy}
\alias{textillateStartProxy}
\alias{textillateStopProxy}
\alias{textillateInProxy}
\alias{textillateOutProxy}
\title{Proxies}
\usage{
textillateStartProxy(proxy)

textillateStopProxy(proxy)

textillateInProxy(proxy)

textillateOutProxy(proxy)
}
\arguments{
\item{proxy}{an object of class \code{txtProxy} as returned by \code{\link{textillateProxy}}.}
}
\description{
Shiny proxies.
}
\examples{
if(interactive()){
  library(shiny)

  ui <- fluidPage(
    actionButton(
      "start",
      "start"
    ),
    actionButton(
      "stop",
      "stop"
    ),
    actionButton(
      "in",
      "in"
    ),
    actionButton(
      "out",
      "out"
    ),
    textillateOutput('textillate')
  )

  server <- function(input, output){
    output$textillate <- renderTextillate({
      textillate("Click to start", auto.start = FALSE) \%>\%
        textillateIn(effect = "fadeIn") \%>\%
        textillateOut(effect = "bounce")
    })

    observeEvent(input$start, {
      textillateProxy("textillate") \%>\%
        textillateStartProxy()
    })

    observeEvent(input$start, {
      textillateProxy("textillate") \%>\%
        textillateStopProxy()
    })

    observeEvent(input$start, {
      textillateProxy("textillate") \%>\%
        textillateInProxy()
    })

    observeEvent(input$out, {
      textillateProxy("textillate") \%>\%
        textillateOutProxy()
    })
  }

  shinyApp(ui, server)
}

}
