% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textillate.R
\name{textillate}
\alias{textillate}
\title{Textillate}
\usage{
textillate(text, ..., loop = FALSE, min.display.time = 2000,
  initial.delay = 0, auto.start = TRUE, in.effects = list(),
  out.effects = list("hinge"), type = "char", width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{...}{any element.}

\item{loop}{enable looping.}

\item{min.display.time}{sets the minimum display time for each text before it is replaced.}

\item{initial.delay}{sets the initial delay before starting the animation.}

\item{auto.start}{set whether or not to automatically start animating.}

\item{in.effects, out.effects}{custom set of 'in' and 'out' effects. This effects whether or not the character
is shown/hidden before or after an animation.}

\item{width, height}{dimensions.}

\item{elementId}{id of html element.}

\item{in.delay, out.delay}{effect name.}

\item{in.delay.scale}{set the delay factor applied to each consecutive character.}

\item{in.delay, out.delay}{set the delay between each character.}

\item{in.sync, out.sync}{set to true to animate all the characters at the same time.}

\item{in.shuffle, out.shuffle}{randomize the character sequence.}

\item{in.reverse, out.reverse}{reverse the character sequence.}

\item{in.callback, out.callback}{callback that executes once the animation has finished.}
}
\description{
Animate text
}
\details{
Effects:
\itemize{
  \item{\code{}}
}
}
\examples{
textillate("Textillate") # basic

textillate("Effects", `in` = list(effect = "rollIn")) # effects

textillate("Duration and effect", `in` = list(effect = "flipInX"), min.display.time = 5000)

# OR
textillate("Duration and effect", min.display.time = 5000) \%>\%
  textillate_in(
    effect = "flipInX",
    delay = 1000
  )

}
\seealso{
\code{\link{textillate_in}}, \code{\link{textillate_out}}
}
