% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inNout.R
\name{textillateIn}
\alias{textillateIn}
\alias{textillateOut}
\title{Customise effects}
\usage{
textillateIn(t, effect = "fadeInLeftBig", delay.scale = 1.5, delay = 50,
  sync = FALSE, shuffle = FALSE, reverse = FALSE)

textillateOut(t, effect = "hinge", delay.scale = 1.5, delay = 50,
  sync = FALSE, shuffle = FALSE, reverse = FALSE)
}
\arguments{
\item{t}{and object of class \code{textillate} as returned by \code{\link{textillate}}.}

\item{effect}{set effect name.}

\item{delay.scale}{set the delay factor applied to each consecutive character.}

\item{delay}{set the delay between each character.}

\item{sync}{set to true to animate all the characters at the same time.}

\item{shuffle}{randomize the character sequence.}

\item{reverse}{reverse the character sequence.}
}
\description{
Cutomise in  and out effect.
}
\details{
See \href{textillate.js.org/}{http://textillate.js.org/} for examples.
\code{reserve = TRUE} doesn't make sense with \code{sync = TRUE}.
}
\examples{
# may not work in RStudio viewer, open in browser
textillate("Duration and effect", min.display.time = 5000) \%>\%
  textillateIn(
    effect = "flipInX",
    delay = 1000,
    shuffle = TRUE
  )

}
