% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{txt_start_p}
\alias{txt_start_p}
\alias{txt_stop_p}
\alias{txt_in_p}
\alias{txt_out_p}
\title{Proxies}
\usage{
txt_start_p(proxy)

txt_stop_p(proxy)

txt_in_p(proxy)

txt_out_p(proxy)
}
\description{
Shiny proxies.
}
\examples{
if(interactive()){
  library(shiny)

  ui <- fluidPage(
    actionButton(
      "start",
      "start"
    ),
    actionButton(
      "stop",
      "stop"
    ),
    actionButton(
      "in",
      "in"
    ),
    actionButton(
      "out",
      "out"
    ),
    txtOutput('textillate')
  )

  server <- function(input, output){
    output$textillate <- renderTxt({
      txt("Click to start", auto.start = FALSE)
    })

    observeEvent(input$start, {
      txtProxy("textillate") \%>\%
        txt_start_p()
    })

    observeEvent(input$start, {
      txtProxy("textillate") \%>\%
        txt_stop_p()
    })

    observeEvent(input$start, {
      txtProxy("textillate") \%>\%
        txt_in_p()
    })

    observeEvent(input$out, {
      txtProxy("textillate") \%>\%
        txt_out_p()
    })
  }

  shinyApp(ui, server)
}

}
