% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textillate.R
\name{textillate}
\alias{textillate}
\title{Textillate}
\usage{
textillate(text, ..., loop = FALSE, min.display.time = 2000,
  initial.delay = 0, auto.start = TRUE, in.effects = list(),
  out.effects = list("hinge"), type = "char", width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{text}{text to animate.}

\item{...}{any element.}

\item{loop}{enable looping.}

\item{min.display.time}{sets the minimum display time for each text before it is replaced.}

\item{initial.delay}{sets the initial delay before starting the animation.}

\item{auto.start}{set whether or not to automatically start animating.}

\item{in.effects, out.effects}{custom set of 'in' and 'out' effects. This effects whether or not the character
is shown/hidden before or after an animation.}

\item{type}{set the type of token to animate (available types: \code{char} and \code{word})}

\item{width, height}{dimensions.}

\item{elementId}{id of html element.}
}
\description{
Animate text
}
\details{
Effects:
\itemize{
  \item{\code{}}
}

See \href{textillate.js.org/}{http://textillate.js.org/} for examples.
\code{reserve = TRUE} doesn't make sense with \code{sync = TRUE}.
}
\examples{
# may not work in RStudio viewer, open in browser
textillate("Textillate") # basic

textillate("Effects", `in` = list(effect = "rollIn")) # effects

textillate("Duration and effect", `in` = list(effect = "flipInX"), min.display.time = 5000)

# OR
textillate("Duration and effect", min.display.time = 5000) \%>\%
  textillate_in(
    effect = "flipInX",
    delay = 1000
  )

}
\seealso{
\code{\link{textillate_in}}, \code{\link{textillate_out}}
}
